Feedback messages added

+++ HealthInstitution/Commands/Patient/RemoveAppointmentCommand.cs
@@ -7,6 +7,7 @@ using System.ComponentModel;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using System.Windows;
 
 namespace HealthInstitution.Commands
 {
@@ -19,12 +20,21 @@ namespace HealthInstitution.Commands
         public override void Execute(object? parameter)
         {
             var apt = GlobalStore.ReadObject<Appointment>("ChosenAppointment");
-            _viewModel._appointmentService.Delete(apt.Id);
-            EventBus.FireEvent("PatientAppointments");
-            /*
-            _viewModel.Appointments.Remove(apt);
-            _viewModel.SelectedAppointment = _viewModel.Appointments.First();
-            */
+            if (DateTime.Now.AddDays(1) > apt.StartDate)
+            {
+                MessageBox.Show("You can't remove appointment in last 24h!");
+            }
+            else 
+            {
+                
+                _viewModel._appointmentService.Delete(apt.Id);
+                MessageBox.Show("Appointment deleted Successfully");
+                EventBus.FireEvent("PatientAppointments");
+                /*
+                _viewModel.Appointments.Remove(apt);
+                _viewModel.SelectedAppointment = _viewModel.Appointments.First();
+                */
+            }
         }
     }
 }
